+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v1.0.4] - 2026-06-21
 
-### Fix
+### Fixed
 
 - Fixed `SaveCursorPos` and `RestoreCursorPos` using `CSI` instead of `Escape`.