+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.1.0] - 2026-07-05
+
 ### Added
 
 - `StripRunes` function strips ANSI escape sequences from `[]rune` input type.